==> Lessons/04_Lab_DataWrangling_withKLnotes.Rmd <==
---
title: "Lab 4: Data Wrangling"
author: "Environmental Data Analytics | John Fay and Luana Lima | Developed by Kateri Salk"
date: "Spring 2023"
output: pdf_document
geometry: margin=2.54cm
editor_options: 
  chunk_output_type: console
---

## Objectives
1- Answer questions on M3/A3
2- Answer questions on M4
3- Practice wrangling datasets with dplyr functions

## Set up your session

Today we will work with a dataset from the [North Temperate Lakes Long-Term Ecological Research Station](https://lter.limnology.wisc.edu/about/overview). The NTL-LTER is located in the boreal zone in northern Wisconsin, USA. We will use the [chemical and physical limnology dataset](https://lter.limnology.wisc.edu/content/cascade-project-north-temperate-lakes-lter-core-data-physical-and-chemical-limnology-1984), running from 1984-2016. 

Opening discussion: why might we be interested in long-term observations of temperature, oxygen, and light in lakes?

> Add notes here: Important to know the health of lakes since they are a water resources and how healthy they are as an ecosystem.

```{r, message = FALSE}
#Install packages
install.packages("here") 
library(tidyverse)
library(lubridate)
library(here) #The here package allows for better control of relative paths

#Ensure that "here" points to your project folder
here()

#Read in the data
NTL.phys.data <- read.csv(
  file=here("Data/Raw/NTL-LTER_Lake_ChemistryPhysics_Raw.csv"), 
  stringsAsFactors = TRUE
)

#Show the datatype of the 'sampledate' column
str(NTL.phys.data$sampledate)

#Alternatively, use the tidyverse/dplyr "glimpse" function
glimpse(NTL.phys.data$sampledate)

# Change sampledate values into date objects
NTL.phys.data$sampledate <- mdy(NTL.phys.data$sampledate)
```

---

## Filter

Filtering allows us to choose certain rows (observations) in our dataset.

```{r}
# note the data types of these two columns
class(NTL.phys.data$lakeid)
class(NTL.phys.data$depth)

# dplyr filtering
NTL.phys.data.surface <- filter(NTL.phys.data, depth == 0) #depth is specifying column name; within the function you do not need quotations to specify the name

# Choose multiple conditions to filter
summary(NTL.phys.data$lakename)
NTL.phys.data.PeterPaul <- 
  filter(NTL.phys.data, lakename %in% c("Paul Lake", "Peter Lake"))

# Choose a range of conditions of a numeric or integer variable
summary(NTL.phys.data$daynum)
NTL.phys.data.JunethruOctober <- filter(NTL.phys.data, daynum %in% c(152:304)) #whenever we have 100 elements becacuse it's a range, because it's sequential values, we can use c() to specify first column and last column
```

```{r Exercise 1 & 2}
# Exercise 1: 
# filter NTL.phys.data for the year 1999
# what code do you need to use, based on the class of the variable?
class(NTL.phys.data$year4)
summary(NTL.phys.data$year4)
NTL.phys.data.1999 <- filter(NTL.phys.data, year4 == 1999) #the value on the other side of == signifies what you want to filter for


# Exercise 2: 
# filter NTL.phys.data for Tuesday Lake from 1990 through 1999.
NTL.phys.data.Ex2 <- filter(NTL.phys.data, lakename == "Tuesday Lake" & year4 %in% c(1990:1999)) #use the & to give the "and" condition; the | is the "or" condition; in this case, could change out & for , but still keep the %in%


```
Question: Why don't we filter using row numbers?

> Answer: We don't filter using row numbers because values vary across the row, and we don't know all the values.

---

## Pipes

Pipe is another method to wrangle datasets that looks cleaner and is easier to read.  We designate a pipe with `%>%`. A good way to think about the function of a pipe is with the word "then." 

Let's say we want to take our raw dataset (NTL.phys.data), *then* filter the data for Peter and Paul lakes, *then* select temperature and observation information, and *then* add a column for temperature in Fahrenheit: 

```{r Exercise 3 & 4}
#Example using pipes to wrangle data
NTL.phys.data.processed <- 
  NTL.phys.data %>%
  filter(lakename == "Paul Lake" | lakename == "Peter Lake") %>%
  select(lakename, sampledate:temperature_C) %>%
  mutate(temperature_F = (temperature_C*9/5) + 32)

#Exercise 3: Using a pipe filter NTL.phys.data for Tuesday Lake from 1990 
# through 1999 only for July.
NTL.phys.data.Ex3 <- 
  NTL.phys.data %>%
  filter(lakename == "Tuesday Lake" & year4 %in% c(1990:1999) & month(sampledate) == 7)
# Note: the month() function is from lubridate

# Note: another way to do this is
NTL.phys.data.Ex3 <- 
  NTL.phys.data %>%
  mutate(Month = month(sampledate)) %>% 
  filter(lakename == "Tuesday Lake" & year4 %in% c(1990:1999) & Month == 7)
# Note: Still using the lubridate column and this generates another column you can filter from

# another way to do it
NTL.phys.data.Ex3 <- 
  NTL.phys.data %>%
  mutate(Month = month(sampledate)) %>%
  filter(Month == 7)

# Exercise 4: Using the data from part 3, a pipe, and the summarize () function
# find the mean surface temp (hint: you will need to add another filter for depth ==0). Make sure you eliminate NAs before computing the means
NTL.phys.data.Ex4 <- 
  NTL.phys.data.Ex3 %>%
  filter(depth ==0) %>%
  drop_na(temperature_C) %>% 
  summarize(meantemp = mean(temperature_C)) 
# Note: inside the pipe, use drop_na() over na.omit() because sometimes it doesn't work the same way as the drop_na(). na.omit() doesn't drop observations, but rather is used in tandem with another function because R cannot calculate when there are NAs in the row/column
# Note: the summarize() function will always give you less observations than when you started
```

## Gather and Spread

For gather we will use `pivot_longer` and for spread we will use `pivot_wider`.

```{r Exercise 5 & 6}
#Exercise 5: Gather irradiance data (measured in the water column and measured
#  on the deck of the sampling boat) into one column using pivot_longer. Name
#  the new column holding the irradiance type as "Irradiance_Type", and name the
#  new column holding the irradiance values as "Irradiance_Value".
NTL.phys.data.Ex5 <- 
  NTL.phys.data %>%
  pivot_longer(irradianceWater:irradianceDeck,
               names_to = "Irradiance_Type",
               values_to = "Irradiance_Vaue")
  
  
  #If you don't pipe: pivot_longer(NTL.phys.data, cols = irradianceWater:irradianceDeck, names_to = "Irradiance_Type", values_to = "Irradiance_Value")

#Exercise 6: Spread temperatureC into more than one column based on the depth.
NTL.phys.data.Ex6 <- 
  NTL.phys.data %>%
  pivot_wider(names_from = depth,
              values_from = temperature_C) #%>%
# select()  and then select the temperature values you want
# pivot_longer() is definitely used much more than pivot_wider


```
